requirements.txt
spotipy
PyQt6
python-dotenv